api2

+++ HangoutsDbLibrary/HangoutsServicesLibrary/Services/ActivityService.cs
@@ -30,7 +30,7 @@ namespace HangoutsBusinessLibrary.Services
             }
         }
 
-        public Activity AddActivity(Activity activity)
+        public string AddActivity(Activity activity)
         {
             using (var uow = new UnitOfWork())
             {
@@ -39,7 +39,7 @@ namespace HangoutsBusinessLibrary.Services
                 Group group = groupRepository.GetByID(activity.Group.ID);
                 if (group == null)
                 {
-                    return null;
+                    return "Invalid group ID";
                 }
                 activity.Group = group;
                 Activity existActivity = activityRepository
@@ -47,14 +47,14 @@ namespace HangoutsBusinessLibrary.Services
                     .Where(a => a.Description == activity.Description && a.GroupID == activity.Group.ID)
                     .FirstOrDefault();
                 if (existActivity != null)
-                    return null;
+                    return "There is already an activity with such a description";
                 activityRepository.Insert(activity);
                 uow.SaveChanges();
-                return activity;
+                return "Ok";
             }
         }
 
-        public Activity UpdateActivity(Activity activity, int id)
+        public string UpdateActivity(Activity activity, int id)
         {
             using (var uow = new UnitOfWork())
             {
@@ -62,31 +62,37 @@ namespace HangoutsBusinessLibrary.Services
                 Activity activityToUpdate = activityRepository.GetByID(id);
                 if (activityToUpdate == null)
                 {
-                    return null; 
+                    return "Invalid ID"; 
                 }
                 Activity existActivity = activityRepository.GetAll().Where(a => a.Description == activity.Description).FirstOrDefault();
-                if (existActivity != null)
-                    return null;
+                if (existActivity == null)
+                {
+                    return "There is already an activity with such a description";
+                }
                 activityToUpdate.Description = activity.Description;
                 activityRepository.Edit(activityToUpdate);
                 uow.SaveChanges();
-                return activityToUpdate;
+                return "Ok";
             }
         }
 
-        public Activity DeleteActivity(int id)
+        public string DeleteActivity(int id)
         {
             using (var uow = new UnitOfWork())
             {
+                var planRepository = uow.GetRepository<Plan>();
+                List<Plan> plan = planRepository.GetAll().Where(p => p.ActivityID == id).ToList();
+                if (plan != null)
+                    return "This activity can not be deleted as long as there are plans containing it";
                 var activityRepository = uow.GetRepository<Activity>();
                 Activity activity = activityRepository.GetByID(id);
                 if (activity == null)
                 {
-                    return null;
+                    return "Invalid ID";
                 }
                 activityRepository.Delete(activity);
                 uow.SaveChanges();
-                return activity;
+                return "Ok";
             }
         }
     }
